requirements.txt
# Tobi Core Dependencies
# Install with: pip install -r requirements.txt

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
websockets>=14.0
python-multipart>=0.0.18

# Intelligence: Claude API (primary)
anthropic>=0.52.0

# Voice: Speech-to-Text
useful-moonshine-onnx>=1.0.0
faster-whisper>=1.1.0

# Voice: Text-to-Speech
# kokoro>=0.7.0 (Optional: has dependency conflicts on some systems)
edge-tts>=6.1.0
soundfile>=0.13.0

# Voice: Wake Word
openwakeword>=0.6.0

# Voice: Audio I/O
pyaudio>=0.2.14
numpy>=1.26.0

# Memory
chromadb>=1.5.0

# Web Tools
ddgs>=7.0.0
beautifulsoup4>=4.12.0
playwright>=1.49.0
pycookiecheat>=0.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
